+++ src/main/java/com/bob/admin/adapter/api/request/ReadManagementMembersRequest.java
@@ -1,0 +1,5 @@
+package com.bob.admin.adapter.api.request;
+
+public record ReadManagementMembersRequest(String key, String keyword) {
+
+}
