+++ src/main/java/com/bob/admin/application/dto/command/ChangeManagementMemberStatusCommand.java
@@ -1,0 +1,7 @@
+package com.bob.admin.application.dto.command;
+
+import java.util.UUID;
+
+public record ChangeManagementMemberStatusCommand(UUID memberId, String status, String memo) {
+
+}
