+++ src/main/java/com/bob/admin/application/port/in/ManagementMemberReader.java
@@ -1,0 +1,15 @@
+package com.bob.admin.application.port.in;
+
+import java.util.UUID;
+
+import org.springframework.data.domain.Pageable;
+
+import com.bob.admin.application.dto.result.ManagementMemberDetail;
+import com.bob.admin.application.port.result.ManagementMemberSummaries;
+
+public interface ManagementMemberReader {
+
+    ManagementMemberSummaries readAll(String key, String keyword, Pageable pageable);
+
+    ManagementMemberDetail readDetail(UUID memberId);
+}
