+++ src/main/java/com/bob/admin/application/port/result/ManagementMemberActivity.java
@@ -1,0 +1,8 @@
+package com.bob.admin.application.port.result;
+
+import com.bob.admin.application.port.result.activity.ManagementMemberPost;
+import com.bob.admin.application.port.result.activity.ManagementMemberTrade;
+
+public record ManagementMemberActivity(ManagementMemberPost post, ManagementMemberTrade trade) {
+
+}
