+++ src/main/java/com/bob/admin/application/port/result/ManagementMemberSummaries.java
@@ -1,0 +1,7 @@
+package com.bob.admin.application.port.result;
+
+import java.util.List;
+
+public record ManagementMemberSummaries(Long totalCount, List<ManagementMember> members) {
+
+}
