+++ src/main/java/com/bob/admin/application/port/result/activity/ManagementMemberPost.java
@@ -1,0 +1,7 @@
+package com.bob.admin.application.port.result.activity;
+
+import java.util.List;
+
+public record ManagementMemberPost(Integer count, List<Long> written) {
+
+}
