+++ src/main/java/com/bob/admin/application/port/result/activity/ManagementMemberTrade.java
@@ -1,0 +1,7 @@
+package com.bob.admin.application.port.result.activity;
+
+import java.util.List;
+
+public record ManagementMemberTrade(Long count, List<Long> sold, List<Long> bought) {
+
+}
