--- src/main/java/com/bob/core/management/adapter/api/request/ReadManagementMembersRequest.java
@@ -1,5 +1,0 @@
-package com.bob.core.management.adapter.api.request;
-
-public record ReadManagementMembersRequest(String key, String keyword) {
-
-}
