--- src/main/java/com/bob/core/management/application/dto/command/ChangeManagementMemberStatusCommand.java
@@ -1,7 +1,0 @@
-package com.bob.core.management.application.dto.command;
-
-import java.util.UUID;
-
-public record ChangeManagementMemberStatusCommand(UUID memberId, String status, String memo) {
-
-}
