--- src/main/java/com/bob/core/management/application/port/result/ManagementMemberActivity.java
@@ -1,8 +1,0 @@
-package com.bob.core.management.application.port.result;
-
-import com.bob.core.management.application.port.result.activity.ManagementMemberPost;
-import com.bob.core.management.application.port.result.activity.ManagementMemberTrade;
-
-public record ManagementMemberActivity(ManagementMemberPost post, ManagementMemberTrade trade) {
-
-}
