--- src/main/java/com/bob/core/management/application/port/result/ManagementMemberSummaries.java
@@ -1,7 +1,0 @@
-package com.bob.core.management.application.port.result;
-
-import java.util.List;
-
-public record ManagementMemberSummaries(Long totalCount, List<ManagementMember> members) {
-
-}
