--- src/main/java/com/bob/core/management/application/port/result/activity/ManagementMemberPost.java
@@ -1,7 +1,0 @@
-package com.bob.core.management.application.port.result.activity;
-
-import java.util.List;
-
-public record ManagementMemberPost(Integer count, List<Long> written) {
-
-}
