--- src/main/java/com/bob/core/management/application/port/result/activity/ManagementMemberTrade.java
@@ -1,7 +1,0 @@
-package com.bob.core.management.application.port.result.activity;
-
-import java.util.List;
-
-public record ManagementMemberTrade(Long count, List<Long> sold, List<Long> bought) {
-
-}
